requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
# created-by: conda 24.11.3
aiohappyeyeballs=2.4.4=py313hecd8cb5_0
aiohttp=3.11.10=py313h46256e1_0
aiosignal=1.2.0=pyhd3eb1b0_0
annotated-types=0.6.0=py313hecd8cb5_0
anyio=4.11.0=pyhcf101f3_0
aom=3.6.0=hcec6c5f_0
appnope=0.1.4=py313hecd8cb5_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py313h46256e1_1
asttokens=3.0.0=py313hecd8cb5_0
async-lru=2.0.4=py313hecd8cb5_0
attrs=24.3.0=py313hecd8cb5_0
av=16.0.1=pypi_0
babel=2.16.0=py313hecd8cb5_0
backoff=2.2.1=py313hecd8cb5_1
beautifulsoup4=4.13.4=py313hecd8cb5_0
blas=1.0=openblas
bleach=6.2.0=py313hecd8cb5_0
bottleneck=1.4.2=py313ha0efc16_0
brotli-python=1.0.9=py313h6d0c2b6_9
brotlicffi=1.0.9.2=py313h6d0c2b6_1
bzip2=1.0.8=h6c40b1e_6
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2025.7.15=hecd8cb5_0
cachetools=5.5.1=py313hecd8cb5_0
cairo=1.16.0=h08824b9_6
certifi=2025.8.3=py313hecd8cb5_0
cffi=1.17.1=py313h9205ec4_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
comm=0.2.1=py313hecd8cb5_0
contourpy=1.3.1=py313h1962661_0
cryptography=44.0.1=py313he15b966_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=hd1e62a0_3
dav1d=1.2.1=h6c40b1e_0
debugpy=1.8.11=py313h6d0c2b6_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py313hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.7.1=h6d0c2b6_0
fontconfig=2.14.1=h269ac6a_3
fonttools=4.55.3=py313h46256e1_0
freetype=2.13.3=h02243ff_0
fribidi=1.0.10=haf1e3a3_0
frozenlist=1.5.0=py313h46256e1_0
gettext=0.21.0=h4e8c18a_2
glib=2.84.2=h918fcf5_0
glib-tools=2.84.2=h918fcf5_0
google-auth=2.38.0=py313hecd8cb5_1
google-genai=1.51.0=pyhcf101f3_0
graphite2=1.3.14=he9d5cce_1
gst-plugins-base=1.14.1=hcec6c5f_1
gstreamer=1.14.1=h6c40b1e_1
h11=0.16.0=py313hecd8cb5_0
harfbuzz=10.2.0=h060d35a_1
httpcore=1.0.9=py313hecd8cb5_0
httpx=0.28.1=py313hecd8cb5_0
icu=73.1=hcec6c5f_0
idna=3.7=py313hecd8cb5_0
ipykernel=6.29.5=py313hecd8cb5_1
ipython=9.1.0=py313hecd8cb5_0
ipython_pygments_lexers=1.1.1=py313hecd8cb5_0
ipywidgets=8.1.5=py313hecd8cb5_0
jedi=0.19.2=py313hecd8cb5_0
jinja2=3.1.6=py313hecd8cb5_0
jpeg=9e=h46256e1_3
json5=0.9.25=py313hecd8cb5_0
jsonschema=4.25.0=py313hecd8cb5_0
jsonschema-specifications=2023.7.1=py313hecd8cb5_0
jupyter=1.1.1=py313hecd8cb5_0
jupyter-lsp=2.2.5=py313hecd8cb5_0
jupyter_client=8.6.3=py313hecd8cb5_0
jupyter_console=6.6.3=py313hecd8cb5_1
jupyter_core=5.8.1=py313hecd8cb5_0
jupyter_events=0.12.0=py313hecd8cb5_0
jupyter_server=2.16.0=py313hecd8cb5_0
jupyter_server_terminals=0.5.3=py313hecd8cb5_0
jupyterlab=4.4.4=py313hecd8cb5_0
jupyterlab_pygments=0.3.0=py313hecd8cb5_0
jupyterlab_server=2.27.3=py313hecd8cb5_0
jupyterlab_widgets=3.0.15=py313hecd8cb5_0
kiwisolver=1.4.8=py313h6d0c2b6_0
krb5=1.20.1=h428f121_1
lcms2=2.16=h31d93a5_1
lerc=4.0.0=h6d0c2b6_0
libabseil=20250127.0=cxx17_h548131f_0
libavif=1.1.1=h46256e1_0
libclang=14.0.6=default_h182e8cd_2
libclang13=14.0.6=default_h9f4243d_2
libcurl=8.14.1=h1448931_1
libcxx=19.1.7=haebbb44_3
libdeflate=1.22=h46256e1_0
libedit=3.1.20230828=h6c40b1e_0
libev=4.33=h9ed2024_1
libevent=2.1.12=h04015c4_1
libffi=3.4.4=hecd8cb5_1
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libglib=2.84.2=heb90364_0
libiconv=1.16=h6c40b1e_3
libkrb5=1.21.3=h54be378_4
libllvm14=14.0.6=h26321d7_4
libmpdec=4.0.0=h46256e1_0
libnghttp2=1.57.0=h9beae6a_0
libopenblas=0.3.29=h3f8574a_0
libpng=1.6.39=h6c40b1e_0
libpq=17.4=h40f65ae_2
libprotobuf=5.29.3=h66a43dd_1
libsodium=1.0.18=h1de35cc_0
libssh2=1.11.1=h3a17b82_0
libtiff=4.7.0=h2dfa3ea_0
libwebp-base=1.3.2=h46256e1_1
libxml2=2.13.8=h6070cd6_0
llvm-openmp=19.1.7=h5048363_2
lmdb=0.9.31=hc8b0dd6_0
lz4-c=1.9.4=hcec6c5f_1
markupsafe=3.0.2=py313h46256e1_0
matplotlib=3.10.0=py313hecd8cb5_1
matplotlib-base=3.10.0=py313hae83f4b_0
matplotlib-inline=0.1.6=py313hecd8cb5_0
mistune=3.1.2=py313hecd8cb5_0
multidict=6.1.0=py313h46256e1_0
mysql=8.4.0=h5364baf_2
nbclient=0.10.2=py313hecd8cb5_0
nbconvert=7.16.6=py313hecd8cb5_0
nbconvert-core=7.16.6=py313hecd8cb5_0
nbconvert-pandoc=7.16.6=py313hecd8cb5_0
nbformat=5.10.4=py313hecd8cb5_0
ncurses=6.5=h923df54_0
nest-asyncio=1.6.0=py313hecd8cb5_0
notebook=7.4.4=py313hecd8cb5_0
notebook-shim=0.2.4=py313hecd8cb5_0
numexpr=2.11.0=py313hc046980_0
numpy=2.3.1=py313h1a2ec72_0
numpy-base=2.3.1=py313hce0bb57_0
openjpeg=2.5.2=h2d09ccc_1
openldap=2.6.10=h5897940_0
openssl=3.0.17=hee2dfae_0
overrides=7.4.0=py313hecd8cb5_0
packaging=25.0=py313hecd8cb5_0
pandas=2.3.1=py313h19bd442_0
pandoc=2.12=hecd8cb5_3
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.4=py313hecd8cb5_0
pcre2=10.42=h9b97e30_1
pexpect=4.9.0=py313hecd8cb5_0
pillow=11.3.0=py313h25d8182_0
pip=25.3=pyhc872135_0
pixman=0.46.4=hb254d66_0
platformdirs=4.3.7=py313hecd8cb5_0
ply=3.11=py313hecd8cb5_1
prometheus_client=0.21.1=py313hecd8cb5_0
prompt-toolkit=3.0.52=pyha770c72_0
prompt_toolkit=3.0.52=hd3eb1b0_1
propcache=0.3.1=py313h46256e1_0
psutil=5.9.0=py313h46256e1_1
ptyprocess=0.7.0=pyhd3eb1b0_3
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.6.1=py313hecd8cb5_0
pyasn1-modules=0.4.2=py313hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.11.7=py313hecd8cb5_0
pydantic-core=2.33.2=py313hb0c8e1d_0
pygments=2.19.1=py313hecd8cb5_0
pyopenssl=25.0.0=py313h8d365df_0
pyparsing=3.2.0=py313hecd8cb5_0
pyqt=5.15.10=py313h6d0c2b6_1
pyqt5-sip=12.13.0=py313h46256e1_1
pysocks=1.7.1=py313hecd8cb5_0
python=3.13.5=h81a7116_100_cp313
python-dateutil=2.9.0post0=py313hecd8cb5_2
python-fastjsonschema=2.20.0=py313hecd8cb5_0
python-json-logger=3.2.1=py313hecd8cb5_0
python-tzdata=2025.2=pyhd3eb1b0_0
python_abi=3.13=0_cp313
pytz=2025.2=py313hecd8cb5_0
pyyaml=6.0.2=py313h46256e1_0
pyzmq=26.2.0=py313h6d0c2b6_0
qt-main=5.15.2=hf83fbd5_12
qtconsole=5.6.1=py313hecd8cb5_1
qtpy=2.4.1=py313hecd8cb5_0
readline=8.3=h49f2429_0
referencing=0.30.2=py313hecd8cb5_0
requests=2.32.4=py313hecd8cb5_0
rfc3339-validator=0.1.4=py313hecd8cb5_0
rfc3986-validator=0.1.1=py313hecd8cb5_0
rpds-py=0.22.3=py313h83de92b_0
rsa=4.7.2=pyhd3eb1b0_1
ruff=0.14.5=pypi_0
seaborn=0.13.2=py313hecd8cb5_3
send2trash=1.8.2=py313hecd8cb5_1
setuptools=72.1.0=py313hecd8cb5_0
sip=6.7.12=py313h6d0c2b6_1
six=1.17.0=py313hecd8cb5_0
sniffio=1.3.0=py313hecd8cb5_0
soupsieve=2.5=py313hecd8cb5_0
sqlite=3.50.2=hc8b0dd6_1
stack_data=0.2.0=pyhd3eb1b0_0
tenacity=9.0.0=py313hecd8cb5_0
terminado=0.17.1=py313hecd8cb5_0
tinycss2=1.4.0=py313hecd8cb5_0
tk=8.6.15=h3a5a201_0
tornado=6.5.1=py313h46256e1_0
tqdm=4.67.1=py313h653c010_0
traitlets=5.14.3=py313hecd8cb5_0
typing-extensions=4.12.2=py313hecd8cb5_0
typing-inspection=0.4.0=py313hecd8cb5_0
typing_extensions=4.12.2=py313hecd8cb5_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=py313hecd8cb5_0
wcwidth=0.2.13=py313hecd8cb5_0
webencodings=0.5.1=py313hecd8cb5_2
websocket-client=1.8.0=py313hecd8cb5_0
websockets=15.0.1=py313h46256e1_0
wheel=0.45.1=py313hecd8cb5_0
widgetsnbextension=4.0.13=py313hecd8cb5_0
xz=5.6.4=h46256e1_1
yaml=0.2.5=haf1e3a3_0
yarl=1.18.0=py313h46256e1_0
zeromq=4.3.5=hcec6c5f_0
zlib=1.2.13=h4b97444_1
zstd=1.5.6=h138b38a_0
